Create the toggle page prototype
This page "toggles" a record (completing an ongoing record
or creating a new one if there is no record ongoing) and
redirects the user to the root page. The toggle page can
be accessed via the /toggle url

The idea is that the user can use this url to rapdly
start/stop a record.

It is a prototype as it is like an experiment, to test if
this is really useful. If it's proven to be, improvements
can be made (like replace the word "toggle" for a better one
and maybe render something nice instead of redirecting the
user to the root page.

Signed-off-by: Felipe Milani <felmilani@gmail.com>