Remove custom time input from Datepicker
UX over UI. Using a custom time input had
a bug: the input would lose focus as soon
as the user type anything, making it hard
if not impossible to use it.

So, this commit reverts to time input to
the default of the react-datepicker library,
which is the default input of the browser.
It is not as pretty as chakra's input but
at least it is usable