start recording when navigating from root page to /now
the link to /now looks is a play button so the user kinda
expect to start recording when clicking it.

there was big problems with the fact that react strict mode
runs effects twice, making the navigation to /now?autostart
create 2 records. the previous commit kinda solved this by
making it impossible to have 2 records with the same start
time.